Address PR review comments
- Remove dbg!(err) statements
- Move setup_test_package to TestEnv as a method accepting a path
